+++ etc/requirements.in
@@ -17,12 +17,11 @@ multipledispatch>=0.6.0
 toolz>=0.8.2
 
 # Scientific Libraries
-numpy>=1.11.3
-pandas>=0.18.1,<=0.22
-#pandas==0.24.0
-#pandas>=1
+numpy==1.19.4
+pandas==1.1.5
 
-pandas-datareader>=0.2.1,<0.9.0
+
+pandas-datareader>=0.9.0
 scipy>=0.17.1
 # Needed for parts of pandas.stats
 patsy>=0.4.0
@@ -61,7 +60,7 @@ bottleneck>=1.0.0
 # Graph algorithms used by zipline.pipeline
 networkx>=1.9.1,<2.0
 # NumericalExpression pipeline terms.
-numexpr>=2.6.1
+numexpr==2.7.1
 
 # Currency Codes
 iso4217>=1.6.20180829
