+++ tests/pipeline/test_blaze.py
@@ -60,13 +60,13 @@ dtypeof = op.attrgetter('dtype')
 asset_infos = (
     (make_simple_equity_info(
         tuple(map(ord, 'ABC')),
-        pd.Timestamp(0),
-        pd.Timestamp('2015'),
+        pd.Timestamp(0, tz='utc'),
+        pd.Timestamp('2015', tz='utc'),
     ),),
     (make_simple_equity_info(
         tuple(map(ord, 'ABCD')),
-        pd.Timestamp(0),
-        pd.Timestamp('2015'),
+        pd.Timestamp(0, tz='utc'),
+        pd.Timestamp('2015', tz='utc'),
     ),),
 )
 simple_asset_info = asset_infos[0][0]
@@ -84,14 +84,15 @@ def with_ignore_sid():
 
 @unittest.skipIf(platform.system() == 'Windows', "Don't run test on windows")
 class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
-    START_DATE = pd.Timestamp(0)
-    END_DATE = pd.Timestamp('2015')
+    START_DATE = pd.Timestamp(0, tz='utc')
+    END_DATE = pd.Timestamp('2015', tz='utc')
 
     @classmethod
     @unittest.skipIf(platform.system() == 'Windows', "Don't run test on windows")
     def init_class_fixtures(cls):
         super(BlazeToPipelineTestCase, cls).init_class_fixtures()
-        cls.dates = dates = pd.date_range('2014-01-01', '2014-01-03')
+        cls.dates = dates = pd.date_range('2014-01-01', '2014-01-03', tz='utc')
+
         cls.asof_dates = asof_dates = dates - pd.Timedelta(days=1)
         cls.timestamps = timestamps = dates - pd.Timedelta(hours=1)
 
@@ -930,10 +931,11 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
 
         df = self.df.copy()
         df['timestamp'] = (
-            pd.DatetimeIndex(df['timestamp'], tz='EST') +
-            timedelta(hours=8, minutes=44)
-        ).tz_convert('utc').tz_localize(None)
-        df.ix[3:5, 'timestamp'] = pd.Timestamp('2014-01-01 13:45')
+            pd.DatetimeIndex(df['timestamp']) +
+            timedelta(hours=14, minutes=44)
+        )
+        df.loc[3:5, 'timestamp'] = pd.Timestamp('2014-01-01 13:45', tz='utc')
+
         expr = bz.data(df, name='expr', dshape=self.dshape)
         loader = BlazeLoader()
         ds = from_blaze(
@@ -959,12 +961,14 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         expected['timestamp'] = expected['timestamp'].dt.normalize().astype(
             'datetime64[ns]',
         ).dt.tz_localize('utc')
-        expected.ix[3:5, 'timestamp'] += timedelta(days=1)
+        expected.loc[3:5, 'timestamp'] += timedelta(days=1)
         expected.set_index(['timestamp', 'sid'], inplace=True)
         expected.index = pd.MultiIndex.from_product((
             expected.index.levels[0],
             self.asset_finder.retrieve_all(expected.index.levels[1]),
         ))
+        expected.index = expected.index.rename(None, level=0)
+
         assert_frame_equal(result, expected, check_dtype=False)
 
     def test_id(self):
@@ -995,7 +999,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         """
         expected = self.df.drop(['timestamp', 'asof_date', 'sid'], axis=1)
         expected.index = pd.MultiIndex.from_product((
-            self.dates.tz_localize('UTC'),
+            self.dates,
             self.asset_finder.retrieve_all(self.asset_finder.sids),
         ))
         self._test_id(
@@ -1037,7 +1041,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             axis=1,
         )
         expected.index = pd.MultiIndex.from_product((
-            self.dates.tz_localize('UTC'),
+            self.dates,
             self.asset_finder.retrieve_all(self.asset_finder.sids),
         ))
         self._test_id(
@@ -1098,7 +1102,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
                       [3, 4]]),
             columns=['other', 'value'],
             index=pd.MultiIndex.from_product(
-                (self.dates.tz_localize('UTC'), self.asset_finder.retrieve_all(
+                (self.dates, self.asset_finder.retrieve_all(
                     self.ASSET_FINDER_EQUITY_SIDS
                 )),
             ),
@@ -1144,7 +1148,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
 
         expected = df.drop(['timestamp', 'asof_date', 'sid'], axis=1)
         expected.index = pd.MultiIndex.from_product((
-            self.dates.tz_localize('UTC'),
+            self.dates,
             self.asset_finder.retrieve_all(self.asset_finder.sids),
         ))
         self._test_id(
@@ -1213,7 +1217,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
                   [0, 1],
                   [0, 1]],
             columns=['other', 'value'],
-            index=self.dates.tz_localize('UTC'),
+            index=self.dates,
         )
         self._test_id_macro(
             df,
@@ -1264,24 +1268,24 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         df = pd.DataFrame(
             columns=['asof_date',        'timestamp', 'sid', 'other', 'value'],
             data=[
-                [T('2013-12-31'), T('2013-12-31 22'),    65,        0,      0],
-                [T('2013-12-31'), T('2013-12-31 23'),    65,        1, np.nan],
-                [T('2013-12-31'), T('2013-12-31 22'),    66,   np.nan, np.nan],
-                [T('2013-12-31'), T('2013-12-31 23'),    66,   np.nan,      1],
-                [T('2013-12-31'), T('2013-12-31 22'),    67,        2, np.nan],
-                [T('2013-12-31'), T('2013-12-31 23'),    67,   np.nan, np.nan],
-                [T('2014-01-01'), T('2014-01-01 22'),    65,   np.nan, np.nan],
-                [T('2014-01-01'), T('2014-01-01 23'),    65,   np.nan,      1],
-                [T('2014-01-01'), T('2014-01-01 22'),    66,   np.nan, np.nan],
-                [T('2014-01-01'), T('2014-01-01 23'),    66,        2, np.nan],
-                [T('2014-01-01'), T('2014-01-01 22'),    67,        3,      3],
-                [T('2014-01-01'), T('2014-01-01 23'),    67,        3,      3],
-                [T('2014-01-02'), T('2014-01-02 22'),    65,        2, np.nan],
-                [T('2014-01-02'), T('2014-01-02 23'),    65,        2, np.nan],
-                [T('2014-01-02'), T('2014-01-02 22'),    66,        3,      3],
-                [T('2014-01-02'), T('2014-01-02 23'),    66,   np.nan, np.nan],
-                [T('2014-01-02'), T('2014-01-02 22'),    67,   np.nan, np.nan],
-                [T('2014-01-02'), T('2014-01-02 23'),    67,   np.nan,      4],
+                [T('2013-12-31', tz='utc'), T('2013-12-31 22', tz='utc'),    65,        0,      0],
+                [T('2013-12-31', tz='utc'), T('2013-12-31 23', tz='utc'),    65,        1, np.nan],
+                [T('2013-12-31', tz='utc'), T('2013-12-31 22', tz='utc'),    66,   np.nan, np.nan],
+                [T('2013-12-31', tz='utc'), T('2013-12-31 23', tz='utc'),    66,   np.nan,      1],
+                [T('2013-12-31', tz='utc'), T('2013-12-31 22', tz='utc'),    67,        2, np.nan],
+                [T('2013-12-31', tz='utc'), T('2013-12-31 23', tz='utc'),    67,   np.nan, np.nan],
+                [T('2014-01-01', tz='utc'), T('2014-01-01 22', tz='utc'),    65,   np.nan, np.nan],
+                [T('2014-01-01', tz='utc'), T('2014-01-01 23', tz='utc'),    65,   np.nan,      1],
+                [T('2014-01-01', tz='utc'), T('2014-01-01 22', tz='utc'),    66,   np.nan, np.nan],
+                [T('2014-01-01', tz='utc'), T('2014-01-01 23', tz='utc'),    66,        2, np.nan],
+                [T('2014-01-01', tz='utc'), T('2014-01-01 22', tz='utc'),    67,        3,      3],
+                [T('2014-01-01', tz='utc'), T('2014-01-01 23', tz='utc'),    67,        3,      3],
+                [T('2014-01-02', tz='utc'), T('2014-01-02 22', tz='utc'),    65,        2, np.nan],
+                [T('2014-01-02', tz='utc'), T('2014-01-02 23', tz='utc'),    65,        2, np.nan],
+                [T('2014-01-02', tz='utc'), T('2014-01-02 22', tz='utc'),    66,        3,      3],
+                [T('2014-01-02', tz='utc'), T('2014-01-02 23', tz='utc'),    66,   np.nan, np.nan],
+                [T('2014-01-02', tz='utc'), T('2014-01-02 22', tz='utc'),    67,   np.nan, np.nan],
+                [T('2014-01-02', tz='utc'), T('2014-01-02 23', tz='utc'),    67,   np.nan,      4],
             ],
         )
         fields = OrderedDict(self.dshape.measure.fields)
@@ -1301,7 +1305,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
                 [3,           4],             # Equity(67 [C])
             ],
             index=pd.MultiIndex.from_product(
-                (self.dates.tz_localize('UTC'), self.asset_finder.retrieve_all(
+                (self.dates, self.asset_finder.retrieve_all(
                     self.ASSET_FINDER_EQUITY_SIDS
                 )),
             ),
@@ -1327,12 +1331,12 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         df = pd.DataFrame(
             columns=['asof_date',        'timestamp', 'other', 'value'],
             data=[
-                [T('2013-12-31'), T('2013-12-31 01'),   np.nan,      1],
-                [T('2013-12-31'), T('2013-12-31 02'),   np.nan,      2],
-                [T('2014-01-01'), T('2014-01-01 01'),        1, np.nan],
-                [T('2014-01-01'), T('2014-01-01 02'),   np.nan,      3],
-                [T('2014-01-02'), T('2014-01-02 01'),        2, np.nan],
-                [T('2014-01-02'), T('2014-01-02 02'),   np.nan, np.nan],
+                [T('2013-12-31', tz='utc'), T('2013-12-31 01', tz='utc'),   np.nan,      1],
+                [T('2013-12-31', tz='utc'), T('2013-12-31 02', tz='utc'),   np.nan,      2],
+                [T('2014-01-01', tz='utc'), T('2014-01-01 01', tz='utc'),        1, np.nan],
+                [T('2014-01-01', tz='utc'), T('2014-01-01 02', tz='utc'),   np.nan,      3],
+                [T('2014-01-02', tz='utc'), T('2014-01-02 01', tz='utc'),        2, np.nan],
+                [T('2014-01-02', tz='utc'), T('2014-01-02 02', tz='utc'),   np.nan, np.nan],
             ],
         )
         fields = OrderedDict(self.macro_dshape.measure.fields)
@@ -1496,6 +1500,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             ('asof_date', 'datetime64[ns]'),
             ('timestamp', 'datetime64[ns]'),
         )
+        df['asof_date'] = pd.to_datetime(df['asof_date'], utc=True)
+        df['timestamp'] = pd.to_datetime(df['timestamp'], utc=True)
         expr = bz.data(df, name='expr', dshape=self.dshape)
 
         T = pd.Timestamp
@@ -1509,16 +1515,16 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         deltas_df_single_sid = pd.DataFrame({
             'value': [0.0, 1.0, 2.0, 3.0],
             'asof_date': [
-                T('2013-12-01'),
-                T('2013-12-15'),
-                T('2013-12-02'),  # not more recent than the previous day
-                T('2013-12-16'),
+                T('2013-12-01', tz='utc'),
+                T('2013-12-15', tz='utc'),
+                T('2013-12-02', tz='utc'),  # not more recent than the previous day
+                T('2013-12-16', tz='utc'),
             ],
             'timestamp': [
-                T('2014-01-01 23:00'),
-                T('2014-01-02 23:00'),
-                T('2014-01-03 23:00'),
-                T('2014-01-04 23:00'),
+                T('2014-01-01 23:00', tz='utc'),
+                T('2014-01-02 23:00', tz='utc'),
+                T('2014-01-03 23:00', tz='utc'),
+                T('2014-01-04 23:00', tz='utc'),
             ],
         })
         sids = asset_info.index
@@ -1556,7 +1562,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             for k, v in expected_views_single_sid.items()
         }
         with tmp_asset_finder(equities=asset_info) as finder:
-            dates = pd.date_range('2014-01-01', '2014-01-05')
+            dates = pd.date_range('2014-01-01', '2014-01-05', tz='utc')
             self._run_pipeline(
                 expr,
                 deltas,
@@ -1581,6 +1587,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             ('asof_date', 'datetime64[ns]'),
             ('timestamp', 'datetime64[ns]'),
         )
+        df['asof_date'] = pd.to_datetime(df['asof_date'], utc=True)
+        df['timestamp'] = pd.to_datetime(df['timestamp'], utc=True)
         expr = bz.data(df, name='expr', dshape=self.dshape)
 
         T = pd.Timestamp
@@ -1615,25 +1623,25 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             ],
             'asof_date': [
                 # swapped order: second row is before the first
-                T('2014-01-02'),
-                T('2014-01-01'),
+                T('2014-01-02', tz='utc'),
+                T('2014-01-01', tz='utc'),
 
                 # chronological order: second row is after the first
-                T('2014-01-03'),
-                T('2014-01-04'),
+                T('2014-01-03', tz='utc'),
+                T('2014-01-04', tz='utc'),
 
                 # swapped order: second row is before the first
-                T('2014-01-06'),
-                T('2014-01-05'),
+                T('2014-01-06', tz='utc'),
+                T('2014-01-05', tz='utc'),
             ],
             'timestamp': [
                 # we learn about all rows in monotonically increasing order
-                T('2013-01-02 22:00'),
-                T('2014-01-02 23:00'),
-                T('2014-01-04 22:00'),
-                T('2014-01-04 23:00'),
-                T('2014-01-06 22:00'),
-                T('2014-01-06 23:00'),
+                T('2013-01-02 22:00', tz='utc'),
+                T('2014-01-02 23:00', tz='utc'),
+                T('2014-01-04 22:00', tz='utc'),
+                T('2014-01-04 23:00', tz='utc'),
+                T('2014-01-06 22:00', tz='utc'),
+                T('2014-01-06 23:00', tz='utc'),
             ],
         })
         sids = asset_info.index
@@ -1665,7 +1673,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         with tmp_asset_finder(equities=asset_info) as finder:
             # The dates queried are non-contiguous. We have two day groups to
             # capture the two day pairs in the input data.
-            dates = pd.to_datetime(['2014-01-03', '2014-01-05', '2014-01-07'])
+            dates = pd.to_datetime(['2014-01-03', '2014-01-05', '2014-01-07'], utc=True)
             self._run_pipeline(
                 expr=expr,
                 deltas=deltas,
@@ -1793,6 +1801,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             ('asof_date', 'datetime64[ns]'),
             ('timestamp', 'datetime64[ns]'),
         )
+        df['asof_date'] = pd.to_datetime(df['asof_date'], utc=True)
+        df['timestamp'] = pd.to_datetime(df['timestamp'], utc=True)
         expr = bz.data(df, name='expr', dshape=self.macro_dshape)
 
         T = pd.Timestamp
@@ -1806,16 +1816,16 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         deltas_df = pd.DataFrame({
             'value': [0.0, 1.0, 2.0, 3.0],
             'asof_date': [
-                T('2013-12-01'),
-                T('2013-12-15'),
-                T('2013-12-02'),  # not more recent than the previous day
-                T('2013-12-16'),
+                T('2013-12-01', tz='utc'),
+                T('2013-12-15', tz='utc'),
+                T('2013-12-02', tz='utc'),  # not more recent than the previous day
+                T('2013-12-16', tz='utc'),
             ],
             'timestamp': [
-                T('2014-01-01 23:00'),
-                T('2014-01-02 23:00'),
-                T('2014-01-03 23:00'),
-                T('2014-01-04 23:00'),
+                T('2014-01-01 23:00', tz='utc'),
+                T('2014-01-02 23:00', tz='utc'),
+                T('2014-01-03 23:00', tz='utc'),
+                T('2014-01-04 23:00', tz='utc'),
             ],
         })
         deltas = bz.data(deltas_df, name='deltas', dshape=self.macro_dshape)
@@ -1836,7 +1846,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         })
 
         with tmp_asset_finder(equities=simple_asset_info) as finder:
-            dates = pd.date_range('2014-01-01', '2014-01-05')
+            dates = pd.date_range('2014-01-01', '2014-01-05', tz='utc')
             self._run_pipeline(
                 expr,
                 deltas,
@@ -1859,6 +1869,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             ('asof_date', 'datetime64[ns]'),
             ('timestamp', 'datetime64[ns]'),
         )
+        df['asof_date'] = pd.to_datetime(df['asof_date'], utc=True)
+        df['timestamp'] = pd.to_datetime(df['timestamp'], utc=True)
         expr = bz.data(df, name='expr', dshape=self.macro_dshape)
 
         T = pd.Timestamp
@@ -1893,25 +1905,25 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             ],
             'asof_date': [
                 # swapped order: second row is before the first
-                T('2014-01-02'),
-                T('2014-01-01'),
+                T('2014-01-02', tz='utc'),
+                T('2014-01-01', tz='utc'),
 
                 # chronological order: second row is after the first
-                T('2014-01-03'),
-                T('2014-01-04'),
+                T('2014-01-03', tz='utc'),
+                T('2014-01-04', tz='utc'),
 
                 # swapped order: second row is before the first
-                T('2014-01-06'),
-                T('2014-01-05'),
+                T('2014-01-06', tz='utc'),
+                T('2014-01-05', tz='utc'),
             ],
             'timestamp': [
                 # we learn about all rows in monotonically increasing order
-                T('2013-01-02 22:00'),
-                T('2014-01-02 23:00'),
-                T('2014-01-04 22:00'),
-                T('2014-01-04 23:00'),
-                T('2014-01-06 22:00'),
-                T('2014-01-06 23:00'),
+                T('2013-01-02 22:00', tz='utc'),
+                T('2014-01-02 23:00', tz='utc'),
+                T('2014-01-04 22:00', tz='utc'),
+                T('2014-01-04 23:00', tz='utc'),
+                T('2014-01-06 22:00', tz='utc'),
+                T('2014-01-06 23:00', tz='utc'),
             ],
         })
         deltas = bz.data(deltas_df, name='deltas', dshape=self.macro_dshape)
@@ -1926,7 +1938,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         with tmp_asset_finder(equities=simple_asset_info) as finder:
             # The dates queried are non-contiguous. We have two day groups to
             # capture the two day pairs in the input data.
-            dates = pd.to_datetime(['2014-01-03', '2014-01-05', '2014-01-07'])
+            dates = pd.to_datetime(['2014-01-03', '2014-01-05', '2014-01-07'], utc=True)
             self._run_pipeline(
                 expr=expr,
                 deltas=deltas,
@@ -1948,6 +1960,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             ('asof_date', 'datetime64[ns]'),
             ('timestamp', 'datetime64[ns]'),
         )
+        df['asof_date'] = pd.to_datetime(df['asof_date'], utc=True)
+        df['timestamp'] = pd.to_datetime(df['timestamp'], utc=True)
         expr = bz.data(df, name='expr', dshape=self.macro_dshape)
 
         T = pd.Timestamp
@@ -1971,21 +1985,21 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             ],
             'asof_date': [
                 # the first row is for current data
-                T('2014-01-02'),
+                T('2014-01-02', tz='utc'),
 
                 # all other rows are restating the same historical value
-                T('2013-12-01'),
-                T('2013-12-01'),
-                T('2013-12-01'),
-                T('2013-12-01'),
+                T('2013-12-01', tz='utc'),
+                T('2013-12-01', tz='utc'),
+                T('2013-12-01', tz='utc'),
+                T('2013-12-01', tz='utc'),
             ],
             'timestamp': [
                 # we learn about all rows within a single calendar index
-                T('2014-01-02 23:00'),
-                T('2014-01-02 23:01'),
-                T('2014-01-02 23:02'),
-                T('2014-01-02 23:03'),
-                T('2014-01-02 23:04'),
+                T('2014-01-02 23:00', tz='utc'),
+                T('2014-01-02 23:01', tz='utc'),
+                T('2014-01-02 23:02', tz='utc'),
+                T('2014-01-02 23:03', tz='utc'),
+                T('2014-01-02 23:04', tz='utc'),
             ],
         })
         deltas = bz.data(deltas_df, name='deltas', dshape=self.macro_dshape)
@@ -1999,7 +2013,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         with tmp_asset_finder(equities=simple_asset_info) as finder:
             # The dates queried are non-contiguous. We have two day groups to
             # capture the two day pairs in the input data.
-            dates = pd.date_range('2014-01-01', '2014-01-03')
+            dates = pd.date_range('2014-01-01', '2014-01-03', tz='utc')
             self._run_pipeline(
                 expr=expr,
                 deltas=deltas,
@@ -2019,8 +2033,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         import blaze as bz
 
         base_dates = pd.DatetimeIndex([
-            pd.Timestamp('2013-12-31'),
-            pd.Timestamp('2014-01-03')
+            pd.Timestamp('2013-12-31', tz='utc'),
+            pd.Timestamp('2014-01-03', tz='utc')
         ])
         repeated_dates = base_dates.repeat(3)
         baseline = pd.DataFrame({
@@ -2113,8 +2127,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         import blaze as bz
 
         base_dates = pd.DatetimeIndex([
-            pd.Timestamp('2013-12-31'),
-            pd.Timestamp('2014-01-03')
+            pd.Timestamp('2013-12-31', tz='utc'),
+            pd.Timestamp('2014-01-03', tz='utc')
         ])
         baseline = pd.DataFrame({
             'value': (0., 1.),
@@ -2139,11 +2153,11 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         })
 
         cal = pd.DatetimeIndex([
-            pd.Timestamp('2014-01-01'),
-            pd.Timestamp('2014-01-02'),
-            pd.Timestamp('2014-01-03'),
+            pd.Timestamp('2014-01-01', tz='utc'),
+            pd.Timestamp('2014-01-02', tz='utc'),
+            pd.Timestamp('2014-01-03', tz='utc'),
             # omitting the 4th and 5th to simulate a weekend
-            pd.Timestamp('2014-01-06'),
+            pd.Timestamp('2014-01-06', tz='utc'),
         ])
 
         def get_expected_output(expected_views, values, asset_info):
@@ -2175,8 +2189,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
                 compute_fn=op.itemgetter(-1),
             )
 
-    test_checkpoints_dates = pd.date_range('2013-12-31', '2014-01-04')
-    test_checkpoints_expected_view_date = pd.Timestamp('2014-01-03')
+    test_checkpoints_dates = pd.date_range('2013-12-31', '2014-01-04', tz='utc')
+    test_checkpoints_expected_view_date = pd.Timestamp('2014-01-03', tz='utc')
 
     @unittest.skipIf(platform.system() == 'Windows', "Don't run test on windows")
     def _test_checkpoints_macro(self, checkpoints, ffilled_value=-1.0):
@@ -2205,7 +2219,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         })
 
         nassets = len(simple_asset_info)
-        expected_views = keymap(lambda t: t.tz_localize('UTC'), {
+        expected_views = keymap(lambda t: t, {
             self.test_checkpoints_expected_view_date: (
                 np.array([[ffilled_value]])
             ),
@@ -2233,8 +2247,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
                 expected_views,
                 expected_output,
                 finder,
-                calendar=pd.date_range('2014-01-01', '2014-01-04'),
-                start=pd.Timestamp('2014-01-03'),
+                calendar=pd.date_range('2014-01-01', '2014-01-04', tz='utc'),
+                start=pd.Timestamp('2014-01-03', tz='utc'),
                 end=dates[-1],
                 window_length=1,
                 compute_fn=op.itemgetter(-1),
@@ -2268,6 +2282,9 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
             'asof_date': [],
             'timestamp': [],
         })
+
+        empty_checkpoints['asof_date'] = pd.to_datetime(empty_checkpoints['asof_date'], utc=True)
+        empty_checkpoints['timestamp'] = pd.to_datetime(empty_checkpoints['timestamp'], utc=True)
 
         self._test_checkpoints_macro(empty_checkpoints)
 
@@ -2338,7 +2355,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
 
         updated_values = baseline.value.iloc[nassets:]
 
-        expected_views = keymap(partial(pd.Timestamp, tz='UTC'), {
+        expected_views = keymap(partial(pd.Timestamp), {
             self.test_checkpoints_expected_view_date: [ffilled_values],
             self.test_checkpoints_dates[-1]: [updated_values],
         })
@@ -2364,8 +2381,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
                 expected_views,
                 expected_output,
                 finder,
-                calendar=pd.date_range('2014-01-01', '2014-01-04'),
-                start=pd.Timestamp('2014-01-03'),
+                calendar=pd.date_range('2014-01-01', '2014-01-04', tz='utc'),
+                start=pd.Timestamp('2014-01-03', tz='utc'),
                 end=dates[-1],
                 window_length=1,
                 compute_fn=op.itemgetter(-1),
@@ -2375,7 +2392,7 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
     def test_checkpoints(self, checkpoints_ts_fuzz_minutes):
         nassets = len(simple_asset_info)
         ffilled_values = (np.arange(nassets, dtype=np.float64) + 1) * 10
-        dates = pd.Index([pd.Timestamp('2014-01-01')] * nassets)
+        dates = pd.Index([pd.Timestamp('2014-01-01', tz='utc')] * nassets)
         checkpoints = pd.DataFrame({
             'sid': simple_asset_info.index,
             'value': ffilled_values,
@@ -2394,8 +2411,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
         checkpoints = pd.DataFrame({
             'sid': [],
             'value': [],
-            'asof_date': [],
-            'timestamp': [],
+            'asof_date': pd.to_datetime([], utc=True),
+            'timestamp': pd.to_datetime([], utc=True)
         })
 
         self._test_checkpoints(checkpoints)
@@ -2458,8 +2475,8 @@ class BlazeToPipelineTestCase(WithAssetFinder, ZiplineTestCase):
                 # asof-dates are flipped in terms of order so that if we
                 # don't sort on asof-date before getting the last in group,
                 # we will get the wrong result.
-                [T('2014-01-03'), T('2014-01-04 00'), 3, 3],
-                [T('2014-01-02'), T('2014-01-04 00'), 2, 2],
+                [T('2014-01-03', tz='utc'), T('2014-01-04 00', tz='utc'), 3, 3],
+                [T('2014-01-02', tz='utc'), T('2014-01-04 00', tz='utc'), 2, 2],
             ],
         )
         fields = OrderedDict(self.macro_dshape.measure.fields)
