+++ tests/pipeline/test_international_markets.py
@@ -115,8 +115,11 @@ class WithInternationalDailyBarData(zf.WithAssetFinder):
                 assets=assets, calendar=calendar, sessions=sessions,
             ))
 
-            panel = (pd.Panel.from_dict(cls.daily_bar_data[name])
-                     .transpose(2, 1, 0))
+            panel = pd.concat(cls.daily_bar_data[name], axis=0)
+            dict_data = {}
+            for column in panel.columns:
+                dict_data[column] = pd.DataFrame(panel[column]).unstack(level=0)
+                dict_data[column].columns = dict_data[column].columns.droplevel(0)
 
             cls.daily_bar_currency_codes[name] = cls.make_currency_codes(
                 calendar,
@@ -124,7 +127,7 @@ class WithInternationalDailyBarData(zf.WithAssetFinder):
             )
 
             cls.daily_bar_readers[name] = InMemoryDailyBarReader.from_panel(
-                panel,
+                dict_data,
                 calendar,
                 currency_codes=cls.daily_bar_currency_codes[name],
             )
