+++ tests/test_history.py
@@ -1772,7 +1772,7 @@ class DailyEquityHistoryTestCase(WithHistory, zf.ZiplineTestCase):
         # 10 days
 
         # get the first 30 days of 2015
-        jan5 = pd.Timestamp('2015-01-05')
+        jan5 = pd.Timestamp('2015-01-05', tz='utc')
 
         # Regardless of the calendar used for this test, equities will
         # only have data on NYSE sessions.
