+++ zipline/algorithm.py
@@ -700,7 +700,7 @@ class TradingAlgorithm(object):
                 self.risk_report = perf
 
         daily_dts = pd.DatetimeIndex(
-            [p['period_close'] for p in daily_perfs], tz='UTC'
+            [p['period_close'] for p in daily_perfs]
         )
         daily_stats = pd.DataFrame(daily_perfs, index=daily_dts)
         return daily_stats
@@ -1645,7 +1645,7 @@ class TradingAlgorithm(object):
             The new symbol lookup date.
         """
         try:
-            self._symbol_lookup_date = pd.Timestamp(dt, tz='UTC')
+            self._symbol_lookup_date = pd.Timestamp(dt)
         except ValueError:
             raise UnsupportedDatetimeFormat(input=dt,
                                             method='set_symbol_lookup_date')
