+++ zipline/data/benchmarks.py
@@ -39,7 +39,7 @@ def get_benchmark_returns_from_file(filelike):
         filelike,
         index_col=['date'],
         parse_dates=['date'],
-    ).tz_localize('utc')
+    )
 
     if 'return' not in df.columns:
         raise ValueError("The column 'return' not found in the "
