+++ zipline/data/in_memory_daily_bars.py
@@ -56,7 +56,7 @@ class InMemoryDailyBarReader(CurrencyAwareSessionBarReader):
     def from_panel(cls, panel, calendar, currency_codes):
         """Helper for construction from a pandas.Panel.
         """
-        return cls(dict(panel.iteritems()), calendar, currency_codes)
+        return cls(dict(panel), calendar, currency_codes)
 
     @property
     def last_available_dt(self):
